feat: 분석, 추천 아카이빙
