+++ services/agent-service/web/src/app/(shell)/agent-result-shared.ts
@@ -11,3 +11,11 @@ export type AnalysisResult = {
   recommendationError?: string
   forecastGaps?: string[]
 }
+
+/** 브라우저 `localStorage`에 쌓는 분석·추천 결과 스냅샷(날짜별 과거 기록 탭용). */
+export type AnalysisHistorySnapshot = {
+  id: string
+  savedAt: string
+  dateKey: string
+  results: AnalysisResult[]
+}
